Adds sp_cmt_cpp_region_regex option
This option lets the user specify which comments they use as region markers, and so
can have different options (currently only spacing).

This also updates the existing tests for region string processing by ensuring it's a
substring match, and that the default doesn't match "BEGIN" or "END" just anywhere in
the line.