Mark parents of semicolons for function calls and protos
If a semicolon is the next real token after a function call or prototype, then this
marks that semicolon's parent type as a function call or prototype (whichever is
correct).

This will allow consistency across detecting the boundaries of functions, because
class function prototypes already have their semicolons' parents marked.