Use consistent boundaries for function splitting
This makes split_fcn_params() be called using the same rules for boundaries for
function prototypes and calls. Now, both of those always have a trailing semi-colon
considered as part of the function, so that it can be taken into consideration when
choosing split points, and so that the generic line splitting code isn't invoked
sometimes for these cases but not other times.