Rewrite core of split_fcn_params()
I was not able to split this part of the patch series up any more than this. It
amounts to a complete rewrite of the code of the function splitting code, which is
now isolated into split_fcn_params_greedy().

The old algorithm would go through the function chunks and when it needs to split the
line, it will search backwards to find a chunk to split the line on. But it would be
confused by "extra" opening and closing parens from function calls within the
original function call. This would often be worked around, essentially, by
do_code_width() which would repeatedly call split_fcn_params() on later chunks that
were also past the legal width. So if split_fcn_params() did an incomplete job on the
first pass, it would often get many more chances.

This new algorithm goes forward through the chunks, remembers the last splittable
point in the line along the way, and then uses that saved pointer when it needs to
split, adding a newline after that saved point, and then re-starting the scanning
from just after the added newline. It also always goes all the way from the first
opening paren to the matching closing paren.

The other critical yet subtle improvement is that the "newpc" functionality in
do_code_width() (added previously in this patch set) is now used for the first
time. Since split_fcn_params_greedy() definitely processes all chunks in the function
the first time through, split_fcn_params() now returns the last chunk it processed so
that do_code_width() will skip over those chunks, and not call split_line() on them.

The unit test result changes reflect the improvements in function splitting, where
additional parens no longer interfere. The new unit tests check for proper handling
of some edge cases, especially that both final closing parens and semicolons
immediately after them do NOT get broken.